backend/requirements.txt
-i https://pypi.org/simple
alembic==1.16.5; python_version >= '3.9'
aniso8601==10.0.1
bcrypt==4.3.0; python_version >= '3.8'
blinker==1.9.0; python_version >= '3.9'
cachelib==0.13.0; python_version >= '3.8'
certifi==2025.8.3; python_version >= '3.7'
click==8.3.0; python_version >= '3.10'
cloudinary==1.44.1
flask==3.1.2; python_version >= '3.9'
flask-bcrypt==1.0.1
flask-cors==6.0.1; python_version >= '3.9' and python_version < '4.0'
flask-jwt-extended==4.7.1; python_version >= '3.9' and python_version < '4'
flask-mail==0.10.0; python_version >= '3.8'
flask-migrate==4.1.0; python_version >= '3.6'
flask-restful==0.3.10
flask-session==0.8.0; python_version >= '3.8'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.2.4; python_version >= '3.9'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
mako==1.3.10; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==4.0.1; python_version >= '3.9'
msgspec==0.19.0; python_version >= '3.9'
psycopg2-binary==2.9.10; python_version >= '3.8'
pyjwt==2.10.1; python_version >= '3.9'
python-dotenv==1.1.1; python_version >= '3.9'
pytz==2025.2
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
sqlalchemy==2.0.43; python_version >= '3.7'
typing-extensions==4.15.0; python_version >= '3.9'
urllib3==2.5.0; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
